--- components/Sidebar/Navigation/Links/Links.module.css
@@ -1,7 +1,0 @@
-.Links {
-    position: relative;
-    text-align: center;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-}
