--- components/Sidebar/Navigation/Navigation.module.css
@@ -1,8 +1,0 @@
-.Sidebar{
-    /*width: 300px;*/
-    /*height: 100%;*/
-    /*padding-top: 50px;*/
-    /*padding-left: 50px;*/
-    /*background: #9a9a9a;*/
-
-}
