--- components/Text/Text.tsx
@@ -1,13 +1,0 @@
-import react from 'react';
-import * as React from "react";
-
-const Text = () => {
-    return (
-        <div>
-            <link href="https://fonts.googleapis.com/css?family=Roboto&display=swap" rel="stylesheet"/>
-            Hi I'm Chris Lim. I'm a Computer Science Student at the University of Toronto!
-        </div>
-    )
-};
-
-export default Text;
