+++ next.config.js
@@ -1,7 +1,7 @@
 const withSass = require('@zeit/next-sass');
 const withCSS = require("@zeit/next-css");
 module.exports = withCSS(withSass({
-    webpack(config, options) {
+    webpack(config, _) {
         config.module.rules.push(
             {
                 test: /\.svg$/,
