+++ package.json
@@ -14,27 +14,27 @@
   "dependencies": {
     "@zeit/next-css": "^1.0.1",
     "@zeit/next-sass": "^1.0.1",
-    "bootstrap": "^4.4.1",
+    "autoprefixer": "^10.1.0",
     "css-loader": "^3.5.2",
     "eject": "0.0.4",
-    "next": "^9.3.5",
-    "next-images": "^1.4.0",
+    "next": "^10.0.4",
+    "next-images": "^1.6.2",
     "now": "^16.7.3",
+    "postcss": "^8.2.2",
     "prop-types": "^15.7.2",
-    "react": "^16.13.1",
-    "react-bootstrap": "^1.0.0",
-    "react-dom": "^16.13.1",
-    "react-scripts": "^3.4.1",
-    "style-loader": "^1.1.4",
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0",
+    "react-scripts": "^3.4.4",
+    "style-loader": "^1.3.0",
+    "tailwindcss": "^2.0.2",
     "url-loader": "^3.0.0"
   },
   "devDependencies": {
-    "@svgr/webpack": "^5.3.1",
-    "@types/node": "^12.12.35",
+    "@svgr/webpack": "^5.5.0",
+    "@types/node": "^12.19.11",
     "@types/prop-types": "^15.7.3",
-    "@types/react": "^16.9.34",
+    "@types/react": "^16.14.2",
     "file-loader": "^5.1.0",
-    "gh-pages": "^2.2.0",
     "typescript": "^3.8.3"
   }
 }
