--- pages/projects.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-import classes from '../css/Projects.module.css'
-
-const Projects = () => {
-    return (
-        <div className={classes.timelineContainer}>Projects Here</div>
-    )
-};
-export default Projects
