+++ tailwind.config.js
@@ -1,0 +1,11 @@
+module.exports = {
+  purge: ["./pages/*.tsx"],
+  darkMode: false, // or 'media' or 'class'
+  theme: {
+    extend: {},
+  },
+  variants: {
+    extend: {},
+  },
+  plugins: [],
+}
